# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/parivaar/public/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/master/districts/{stateId}": {
            "get": {
                "tags": [
                    "Master"
                ],
                "description": "API to get districts",
                "operationId": "districts",
                "parameters": [
                    {
                        "name": "stateId",
                        "in": "path",
                        "description": "stateId",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/addresses/{address_id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update address",
                "description": "This api will update address.",
                "operationId": "updatead",
                "parameters": [
                    {
                        "name": "address_id",
                        "in": "path",
                        "description": "Id of address",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address_type": {
                                        "type": "number",
                                        "example": "1:Billing,2:Shipping"
                                    },
                                    "is_default": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John wick"
                                    },
                                    "address1": {
                                        "type": "string",
                                        "example": "Wall Street 1"
                                    },
                                    "address2": {
                                        "type": "string",
                                        "example": "Near ATM"
                                    },
                                    "country": {
                                        "type": "number",
                                        "example": 103
                                    },
                                    "state": {
                                        "type": "number",
                                        "example": 21
                                    },
                                    "district": {
                                        "type": "number",
                                        "example": 21
                                    },
                                    "city": {
                                        "type": "number",
                                        "example": 21
                                    },
                                    "pincode": {
                                        "type": "number",
                                        "example": 741258
                                    },
                                    "status": {
                                        "type": "number",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete address",
                "description": "This api will delete address.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "address_id",
                        "in": "path",
                        "description": "Id of address",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}